fix: change board[row][col] to board[col][row]
Example :
  Serial.println("Case A1 value: " + String(board[0][0]));
  Serial.println("Case B5 value: " + String(board[1][4]));
  Serial.println("Case C3 value: " + String(board[2][2]));
  Serial.println();

+++ Code/chessLab/src/io/driver/LedStripe.hpp
@@ -9,7 +9,7 @@ private:
   static Adafruit_NeoPixel strip;
 
   LedStripe();
-  int to_index(int row, int col);
+  uint8_t to_index(uint8_t row, uint8_t col);
 
 public:
   static LedStripe &getInstance();
