+++ Code/chessLab/src/io/hal/HallSensor.hpp
@@ -4,12 +4,12 @@ class HallSensor
 {
 private:
   HallSensor();
-  void select_row(int row);
-  void select_col(int col);
+  void select_row(uint8_t row);
+  void select_col(uint8_t col);
 
 public:
   static HallSensor &getInstance();
-  int read(int row, int col);
+  int read(uint8_t row, uint8_t col);
 };
 
 #define hallSensor HallSensor::getInstance()
